requirements.txt
PySocks==1.6.8
pyTelegramBotAPI==3.6.6
python-dotenv==0.9.1
requests==2.19.1
